pyproject.toml
[tool.poetry]
name = "sgi-api-core"
version = "0.0.1"
description = ""
authors = ["Leonel Gareis <leonelgareis@protonmail.ch>"]
license = "MIT"
readme = "README.md"
packages = [{include = "sgi_api_core"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.3"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
pytest = "^7.1.3"
isort = "^5.10.1"
commitizen = "^2.35.0"
pre-commit = "^2.20.0"
jedi = "^0.18.1"
neovim = "^0.3.1"


[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.1"
tag_format = "$version"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
